refactor(builder): move createScript out of the package barrel
sources/{applications,system,windows}.ts imported createScript from
'../index.js', but createScript was declared in index.ts itself, so each
source module imported the barrel back and formed a cycle:

  index.ts -> sources/index.ts -> sources/<mod>.ts -> index.ts

Move createScript beside the AppleScriptBuilder it constructs. index.ts
already re-exports everything from builder.js, so the public API is
unchanged and no new barrel entry is needed.

Also remove src/bun.ts and src/node.ts. Both were one-line re-export
shims with no importers: the ./bun and ./node package entry points map
to index.bun.ts and index.node.ts, and neither shim appears in
package.json exports or the tsup entry list.

resect audit: 3 cycles -> 0, index.ts fan-in 5 -> 2, 0 orphan files.
