--- src/bun.ts
@@ -1,1 +1,0 @@
-export { runtime } from './runtime/bun.js';
