--- src/node.ts
@@ -1,1 +1,0 @@
-export { runtime } from './runtime/node.js';
